Replace CapErr with io::Error
CapErr simply served to wrap an io::Error around an enum that mainly
just served to tell the user what function he had just called.  But the
user already knew that.  Simpler just to return io::Error directly.  The
only complication was Directory::open_file, which needs to handle a
NulError.  I mapped that to io::Error with an ErrorKind of Other.

Fixes #25