Update the nix dev dependency
It doesn't make a difference for the features that we're currently
using, though.  I'm mainly doing this so we don't start to accidentally
use an older non-I/O Safe API, many of which have been converted in nix
0.28 and 0.29.  But leave the minimum Nix version unchanged, so
downstream users won't have to pull in a new version of Nix if they're
already using an old one.